optimized session, and routing
