+++ .env.default
@@ -4,3 +4,4 @@ FORWARD_DB_PORT=3306
 DB_USERNAME=pwmanager
 DB_PASSWORD=password
 DB_DATABASE=pwmanager
+SECRET=secret
