+++ .gitignore
@@ -1,8 +1,3 @@
-# my env file
-.env.Sven
-
-
-
 # Logs
 logs
 *.log
@@ -107,3 +102,5 @@ dist
 
 # TernJS port file
 .tern-port
+
+sessions/
