+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "start": "nodemon index.js",
+    "start": "nodemon --ignore sessions index.js",
     "migrate": "db-migrate up",
     "migrate:rollback": "db-migrate down",
     "migrate:reset": "db-migrate reset",
@@ -34,7 +34,8 @@
     "moment": "^2.29.4",
     "mongoose": "^6.1.5",
     "mysql2": "^2.3.3",
-    "raw-body": "^2.5.1"
+    "raw-body": "^2.5.1",
+    "session-file-store": "^1.5.0"
   },
   "repository": {
     "type": "git",
