+++ views/index.ejs
@@ -30,7 +30,7 @@
                         <ul class="navbar-nav navbar-nav-right">
                             <li class="nav-item dropdown d-none d-lg-block">
 
-                                <form action="/logout" method="POST">
+                                <form action="/customer/logout" method="POST">
 
                                     <input type="text" name="newtask" placeholder="add new task" hidden>
                                     <button
